third_party/WebKit/LayoutTests/crypto/subtle/aes-cbc/generate-key-expected.txt

Test generating an AES key using AES-CBC algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Generating a key...
error is: TypeError: AesKeyGenParams: length: Missing required property
error is: TypeError: AesKeyGenParams: length: Is not a number
error is: TypeError: AesKeyGenParams: length: Is not a number
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 128
PASS key.usages is ['encrypt', 'decrypt']
PASS successfullyParsed is true

TEST COMPLETE
